+++ src/main/java/org/umc/travlocksserver/domain/member/entity/MemberConsent.java
@@ -1,9 +1,7 @@
-package org.umc.travlocksserver.domain.memberconsent.entity;
+package org.umc.travlocksserver.domain.member.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.umc.travlocksserver.domain.member.entity.Member;
-import org.umc.travlocksserver.domain.policy.entity.Policy;
 import org.umc.travlocksserver.global.entity.CreatedSoftDeleteBaseEntity;
 
 @Getter
