+++ src/main/java/org/umc/travlocksserver/domain/template/entity/TemplateCity.java
@@ -2,7 +2,7 @@ package org.umc.travlocksserver.domain.template.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.umc.travlocksserver.domain.city.entity.City;
+import org.umc.travlocksserver.domain.location.entity.City;
 import org.umc.travlocksserver.global.entity.BaseEntity;
 
 @Getter
