+++ src/main/java/org/umc/travlocksserver/domain/template/entity/TemplateTag.java
@@ -2,7 +2,6 @@ package org.umc.travlocksserver.domain.template.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.umc.travlocksserver.domain.tag.entity.Tag;
 import org.umc.travlocksserver.global.entity.CreatedBaseEntity;
 
 @Getter
